Squashed 'include/shark/LinAlg/BLAS/' changes from 51a4469b..99eb90f7
99eb90f7 added most of the missing math functions
09109a7d fixed typo resulting in wrong iterator type chosen
16f99487 added frontend for eigenvalue decomposition and fixed a few smaller issues
b88384c2 added new rule for product with identity matrix
5aded671 fixed warnings on MSVC

git-subtree-dir: include/shark/LinAlg/BLAS
git-subtree-split: 99eb90f7823769f4839c31bb2abdcb1d70401693

+++ kernels/default/pstrf.hpp
@@ -99,7 +99,7 @@ std::size_t pstrf(
 			//such that j == pivot and Ak(j,j) = pivots(pivot)
 			std::size_t pivot = std::max_element(pivots.begin()+j,pivots.end())-pivots.begin();
 			if(pivot != j){
-				P()(k+j) = pivot+k;
+				P()(k+j) = (int)(pivot+k);
 				A().swap_rows(k+j,k+pivot);
 				A().swap_columns(k+j,k+pivot);
 				std::swap(pivots(j),pivots(pivot));
